Revert "bugFix: 更该getValue精度"
This reverts commit e7e4a4bccf9f8dea7ef4f4ea55de39751b0c50fd.

+++ src/observer/sql/executor/tuple.cpp
@@ -33,9 +33,9 @@ struct filter_map {
  * @return
  */
 double getvalue(Tuple *tuple,TupleSchema *schema,RelAttr &attr,int *success,std::shared_ptr<TupleValue>& tupleValue){
-    double left;
-    double right;
-    double value;
+    int left;
+    int right;
+    int value;
     switch (attr.extype) {
         case E:{
             if(attr.second== nullptr){
